LivrariaLCA/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-braces==1.15.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
